fix: deduplicate resolveWhatsAppPhoneNumber into shared utility
